Ability to interrupt Serve() through a context
We add ServeContext() which acts exactly like Serve() but takes a
context as a first argument. When the context is cancelled,
ServeContext() will return. Also add a test for that (the test could be
expanded to check more stuff).

Note that the user has to close the "conn" after cancelling the context,
otherwise, there would be a leak of a goroutine. Another possible
improvement would be to pass the context to the handler as well.